fix: cast message size to int in GetMessagesFromQueue return value

+++ src/agent/src/message_queue_utils.cpp
@@ -23,7 +23,7 @@ GetMessagesFromQueue(std::shared_ptr<IMultiTypeQueue> multiTypeQueue,
         output += "\n" + message.metaData + "\n" + message.data.dump();
     }
 
-    co_return std::tuple<int, std::string> {messages.size(), output};
+    co_return std::tuple<int, std::string> {static_cast<int>(messages.size()), output};
 }
 
 void PopMessagesFromQueue(std::shared_ptr<IMultiTypeQueue> multiTypeQueue, MessageType messageType, int numMessages)
